Autodetect python indent is now smarter
Watch those assumptions!!

I encountered one source code file with a highly indented formatted string.
This was the first indented non-blank line in the file, so my program dutifully
assumed the shiftwidth should be 38!  I *could* restrict it to common values
such as 2, 3, 4, and 8, but a smarter way is to check the context.

To wit: we *now* assume the shiftwidth is given by the indent of the first
nonblank line following a non-indented nonblank line which ends with a colon.