If no good guesses for sw, don't try guessing sw
Previously, code would spit out a rude error when opening a python buffer which
didn't contain any of the relevant types of block (class, def, if, etc.) which
we might use to "guess" the user's indent.  This is worse than it might sound:
for example, it affects every new python file!  Fixed that by catching the
exception and doing nothing.